requirements.txt
PyGithub>=1.58.0
PyYAML>=5.1
